BE/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
langchain-google-genai
python-dotenv
pydantic[email]
python-multipart
python-socketio
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
cloudinary
google-generativeai
websockets
